requirements.txt
Pillow
Flask
Flask-Cors